gacha-gen: split gacha description text gen to new file
